Convert all monetary fields from decimal to integer format
Backend Changes:
- Update all model fields from DecimalField to IntegerField
- Change Transaction.amount, Envelope.budgeted_amount, SavingsGoal fields
- Update RecurringTransaction.amount to use IntegerField
- Fix all serializers to handle integers instead of decimals
- Update views to use int() instead of Decimal()
- Add missing timezone import to models.py
- Create and run database migration for field changes

Frontend Changes:
- Update currency formatting to show 0 decimal places
- Change formatCurrency to use minimumFractionDigits: 0
- Update TransactionModal amount input to use step='1' and min='1'
- Fix currency utility type conversion for string/number handling

Database Migration:
- Create migration to convert all decimal fields to integer fields
- Successfully apply migration to update database schema
- Maintain data integrity during field type conversion

Result:
- All monetary values now display as integers without .00
- VT 1,000 instead of VT 1,000.00 throughout the app
- Cleaner, more appropriate formatting for Vanuatu Vatu currency
- Consistent integer handling across frontend and backend

+++ backend/tracker/serializers.py
@@ -35,8 +35,8 @@ class CategorySerializer(serializers.ModelSerializer):
 
 class EnvelopeSerializer(serializers.ModelSerializer):
     category_name = serializers.CharField(source='category.name', read_only=True)
-    spent_amount = serializers.DecimalField(max_digits=12, decimal_places=2, read_only=True)
-    remaining_amount = serializers.DecimalField(max_digits=12, decimal_places=2, read_only=True)
+    spent_amount = serializers.IntegerField(read_only=True)
+    remaining_amount = serializers.IntegerField(read_only=True)
     percentage_used = serializers.FloatField(read_only=True)
     is_over_budget = serializers.BooleanField(read_only=True)
     is_near_limit = serializers.BooleanField(read_only=True)
@@ -122,16 +122,16 @@ class TransactionSerializer(serializers.ModelSerializer):
 
 
 class BalanceSerializer(serializers.Serializer):
-    total_income = serializers.DecimalField(max_digits=12, decimal_places=2)
-    total_expenses = serializers.DecimalField(max_digits=12, decimal_places=2)
-    balance = serializers.DecimalField(max_digits=12, decimal_places=2)
-    monthly_income = serializers.DecimalField(max_digits=12, decimal_places=2)
-    monthly_expenses = serializers.DecimalField(max_digits=12, decimal_places=2)
+    total_income = serializers.IntegerField()
+    total_expenses = serializers.IntegerField()
+    balance = serializers.IntegerField()
+    monthly_income = serializers.IntegerField()
+    monthly_expenses = serializers.IntegerField()
 
 
 class SavingsGoalSerializer(serializers.ModelSerializer):
     progress_percentage = serializers.FloatField(read_only=True)
-    remaining_amount = serializers.DecimalField(max_digits=12, decimal_places=2, read_only=True)
+    remaining_amount = serializers.IntegerField(read_only=True)  # Changed from DecimalField to IntegerField
 
     class Meta:
         model = SavingsGoal
